refactor: 충돌 해결

+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/controller/RunnerPostController.java
@@ -185,11 +185,11 @@ public class RunnerPostController {
     ) {
         final Page<RunnerPost> pageRunnerPosts = runnerPostService.readRunnerPostsBySupporterIdAndReviewStatus(pageable, supporter.getId(), reviewStatus);
         final List<RunnerPost> foundRunnerPosts = pageRunnerPosts.getContent();
-        final List<Integer> applicantCounts = collectApplicantCounts(pageRunnerPosts);
+        final List<Long> applicantCounts = collectApplicantCounts(pageRunnerPosts);
         final List<RunnerPostResponse.ReferencedBySupporter> responses = IntStream.range(0, foundRunnerPosts.size())
                 .mapToObj(index -> {
                     final RunnerPost foundRunnerPost = foundRunnerPosts.get(index);
-                    final Integer applicantCount = applicantCounts.get(index);
+                    final Long applicantCount = applicantCounts.get(index);
 
                     return RunnerPostResponse.ReferencedBySupporter.of(foundRunnerPost, applicantCount);
                 }).toList();
@@ -200,7 +200,7 @@ public class RunnerPostController {
         return ResponseEntity.ok(PageResponse.from(pageResponse));
     }
 
-    private List<Integer> collectApplicantCounts(final Page<RunnerPost> pageRunnerPosts) {
+    private List<Long> collectApplicantCounts(final Page<RunnerPost> pageRunnerPosts) {
         final List<Long> runnerPostIds = pageRunnerPosts.stream()
                 .map(RunnerPost::getId)
                 .toList();
